feat: detach device

+++ app/routes/app.devices.$deviceId.remove.tsx
@@ -9,15 +9,23 @@ import { DASHBOARD_PREFIX } from "./app"
 
 type LoaderData = {
   device: Device
+  intent: Intent
 }
+
+type Intent = "detach" | "delete"
 
 export const loader: LoaderFunction = async ({ request, params }) => {
   const token = await getSessionToken(request)
   if (!params.deviceId) throw new Response("Device Not Found", { status: 404 })
+  const url = new URL(request.url)
+  let intent: Intent = (url.searchParams.get("intent") ?? "detach") as Intent
+  if (intent !== "detach" && intent !== "delete") {
+    intent = "detach"
+  }
   return api
     .getDeviceDetails(params.deviceId, token)
     .then(data => {
-      return json<LoaderData>({ device: data as Device })
+      return json<LoaderData>({ device: data as Device, intent })
     })
     .catch(err => {
       throw new Response(
@@ -36,51 +44,88 @@ export const loader: LoaderFunction = async ({ request, params }) => {
 export const action: ActionFunction = async ({ request, params }) => {
   const token = await getSessionToken(request)
   if (!params.deviceId) throw new Response("Device Not Found", { status: 404 })
+  const formData = await request.formData()
+  const intent = formData.get("intent")
+  if (intent === "delete")
+    return api
+      .deleteDevice(params.deviceId, token)
+      .then(() => {
+        return redirect(`${DASHBOARD_PREFIX}/groups/${params.groupId ?? ""}`)
+      })
+      .catch(err => {
+        throw new Response(
+          "Error deleting the device: " + err.response?.data?.message ??
+            err.response?.data ??
+            err.response ??
+            err,
+          {
+            status: err.response?.status ?? 500,
+            statusText:
+              err.response?.statusText ?? "Internal Error Deleting The Device!",
+          }
+        )
+      })
   return api
-    .deleteDevice(params.deviceId, token)
+    .detachDevice(params.deviceId, token)
     .then(() => {
       return redirect(`${DASHBOARD_PREFIX}/groups/${params.groupId ?? ""}`)
     })
     .catch(err => {
       throw new Response(
-        "Error deleting the device: " + err.response?.data?.message ??
+        "Error detaching device from the group: " +
+          err.response?.data?.message ??
           err.response?.data ??
           err.response ??
           err,
         {
           status: err.response?.status ?? 500,
           statusText:
-            err.response?.statusText ?? "Internal Error Deleting The Device!",
+            err.response?.statusText ?? "Internal Error Detaching The Device!",
         }
       )
     })
 }
 
 export default function DeleteDevicePage() {
-  const { device } = useLoaderData<LoaderData>()
+  const { device, intent } = useLoaderData<LoaderData>()
   const navigation = useNavigation()
 
   return (
     <div className="DeleteDeviceCard">
       <div className="card">
         <b>
-          Are you sure you want to delete {device.name}{" "}
-          <code className="text-rose-600">
+          Are you sure you want to {intent} {device.name}{" "}
+          <code
+            className={
+              intent === "delete" ? "text-rose-600" : "text-orange-600"
+            }
+          >
             {device.cpuId || "NO_CPU_ID"} - {device.id}
           </code>
         </b>
         <div className="buttons">
           <Form method="POST">
+            <input
+              type="text"
+              name="intent"
+              value={intent}
+              className="hidden"
+              readOnly
+            />
             <TextButton
               type="submit"
-              className="!bg-rose-500"
+              className={
+                intent === "delete" ? "!bg-rose-500" : "!bg-orange-500"
+              }
               disabled={
                 navigation.state === "submitting" ||
                 navigation.state === "loading"
               }
             >
               {navigation.state === "submitting"
-                ? `Deleting ${device.name}...`
+                ? `${intent === "delete" ? "Deleting" : "Detaching"} ${
+                    device.name
+                  }...`
                 : navigation.state === "loading"
                 ? "Redirecting..."
                 : "Yes, I know what I'm doing!"}
